Introduce Simulation Pre-Processing
This commit introduces Simulation Pre-Processing. A first usage can
be seen in `HoverflyConfigTest`.

The challenge was especially to support JUnit 5, as we can only provide
a class-reference in the annotation and not a ready-instantiated class.
Having this, this commit suggests to circumvent the access control for
class instantiation.

This commit does not yet contain all relevant tests (especially for
JUnit 4). In its current state it is meant as RFC, to get feedback
if the solution matches the Hoverfly-Java architecture. Thus follow-up
commits are expected to increase the test-coverage.

Changes in detail:

hoverfly/junit/api:

* Extended HoverflyClientException by message/cause constructor and
    added recommended serial version UID.

hoverfly/junit/core:

* Introduces SimulationPreprocessor to the configurations (Base, Local, Remote)
* Hoverfly: Extended to apply the preprocessor prior to handing it over
    to the Hoverfly client instance. Required to provide an extended version
    of `simulate` in order to support instantiating nested inner non-static
    classes as it is quite common when you write nested JUnit 5 tests.

New classes:

* SimulationPreprocessor: The central class for simulation preprocessing.
    It assumes a mutable Simulation instance which, according to the
    discussion in SpectoLabs/hoverfly-java#193 is just fine. Note, that
    this feature-dev will not include enhancing customization of
    Simulation instances.
* SimulationPreprocessorProvider: This provider contains the algorithm how
    to create an instance of SimulationPreprocessor for various scenarios.
    It is only required for JUnit 5 support, as we can instantiate the
    preprocessor in JUnit 4 directly.
* UnsetSimulationPreprocessor: Required as default value for JUnit 5, as
    `null` is not a valid default for annotation parameters. Could not
    guess of a better solution.

See Also:

* SpectoLabs/hoverfly-java#193